Rewrite `col_binomial`
The previous implementation of `col_binomial` prevented the user from supplying a base definition where `summarised = TRUE`. Instead, the base count was taken as the length of `n`. This meant that the base for a vector like `c(1, 0, 1, 9, 1)` would be 5, whereas we might want to be able to exclude the 9 from the calculation to have `n=3` and `N=4`.